% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_token}
\alias{get_token}
\title{Get Access Token for Health Planet API.}
\usage{
get_token(user_id = NULL, user_password = NULL)
}
\description{
Get Access Token for Health Planet API. You need to have your own account on [Health Planet](https://www.healthplanet.jp/).
}
