% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getToken}
\alias{getToken}
\title{Get Access Token for Health Planet API.}
\usage{
getToken()
}
\description{
Get Access Token for Health Planet API. You need to have your own account on [Health Planet](https://www.healthplanet.jp/).
}

