% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{innerscan}
\alias{innerscan}
\title{Get the innerscan data}
\usage{
innerscan(token)
}
\arguments{
\item{access_token}{the token given by getToken()}
}
\description{
Get the innerscan data via HelthPlanet API with Access Token.
}
